+++ src/layout/FileUpload/FileUploadComponent.module.css
@@ -1,13 +1,5 @@
-tr.blueUnderline th {
-  border-bottom: 2px solid #1eadf7;
-}
-
 .fileUploadTable > tbody > tr > td {
   padding: 10px 4px 10px 6px;
-}
-.fileUploadTable > thead > tr > th {
-  font-size: 0.875rem;
-  padding: 4px 4px 4px 6px;
 }
 
 .fileUploadTable > tbody > tr > td:first-child {
