+++ src/layout/FileUpload/shared/FileTableHeader.module.css
@@ -1,0 +1,5 @@
+.blueUnderline th {
+  border-bottom: var(--solid-border-primary);
+  font-size: 0.875rem;
+  padding: 4px 4px 4px 6px;
+}
